refactor(codec): add basic_datastream alias name
